requirements.txt

aim-ui==3.25.1 
aim==3.25.1 
aimrecords==0.0.7 
aimrocks==0.5.2
aiofiles==22.1.0 
alembic==1.8.1 
base58==2.0.1 
bottleneck==1.3.7 
cachetools==5.2.0
cmaes==0.10.0 
cffi==1.15.1 
click==8.1.3 
colorama==0.4.5
cramjam==2.6.2 
cryptography==38.0.2 
cycler==0.11.0 
deap==1.3.1 
fastapi==0.69.0
fastparquet==2023.2.0 
filelock==3.8.0 
fonttools==4.34.4 
fortranformat==1.2.0 
fsspec==2023.4.0 
greenlet==1.1.3.post0
grpcio==1.53.0 
h11==0.14.0 
jinja2==3.1.2 
joblib==1.2.0 
kiwisolver==1.4.3 
llvmlite==0.39.1 
mako==1.2.3 
markupsafe==2.1.1 
matplotlib==3.5.2 
numba==0.56.4 
numexpr==2.8.4 
numpy==1.23.1 
packaging==21.3 
pandas==2.0.0 
pandas[parquet]==2.0.0 
patsy==0.5.2 
pillow==9.2.0 
protobuf==3.20.3 
psutil==5.9.1 
py3nvml==0.2.7 
pyarrow==11.0.0 
pycparser==2.21 
pydantic==1.10.2 
pyod==1.0.5 
pyparsing==3.0.9 
python-dateutil==2.8.2 
pytz==2022.1 
pyyaml==6.0 
restrictedpython==5.2 
scienceplots==2.0.1 
scikit-learn==1.2.2 
scipy==1.8.1 
seaborn==0.11.2 
setuptools==67.6.1 
six==1.16.0 
sqlalchemy==1.4.42 
starlette==0.15.0
statsmodels==0.13.2 
tables==3.7.0 
threadpoolctl==3.1.0 
tqdm==4.64.0 
typing-extensions==4.3.0 
tzdata==2023.3 
uvicorn==0.18.3 
xmltodict==0.13.0 